requirements.txt
django >= 3.1.4, < 4.0
django-colorfield >= 0.3.2
django-environ >= 0.4.5, < 0.5
pillow >= 8.0.1
